Add Windows specific wcwidth function
Windows doesn't provide a wcwidth() function nor something similar.
However it is possible to measure the width of character by printing
each character and retriving the cursor position before and after
this action.

The measuring program is currently written in ruby, but it is kept
simple enough to be easily converted to C, if this approach proves
to be suitable.

With this implementation it is now possibly to corretly work with cyrillic
characters for instance.