Windows: Fix reception of 4 byte UTF-16 characters
The console is able to receive 4 byte UTF-16 characters, when a proper
font is selected. Still it's unable to display them (only UTC-2).
Therefore in the output these characters are substituted, but it is
possible this way to enter characters from the full Unicode range.

Use better variable naming, while being over it.